Flesh out some higher-level APIs.
The higher-level IPs generally start with "Get" (e.g., GetIPs,
GetNameServers, GetServices).  These APIs take care of
aggregating one or more queries and parsing the responses into
an easy-to-use format.

For example, GetNameServers perforsm an NS query to get the
nameservers for a given domain.  Sometimes these queries do not
return results, in which case GetNameServers falls back to
gleaning the nameserver information from the SOA record, if
available. Once assembling a list of nameservers, GetNameServers
then makes A and AAAA queries to get the IPs for each name
server.  On success, the final result is a
[]*dnsclient.NameServer, where the NameServer struct has two
fields: Name (for the nameserver's domainname), and Addrs, which
is a slice of netip.Addr.