Simpify configuration, Client interface, and dnsclient cmd.
The configuration is now flat: there is a single Config struct;
I removed the transport-specific structs (Do53Config, DoTConfig,
and DoHConfig).  This ends up being much more elegant.
Additionally, the configuration closly matches the options of
the dig(1) tool, which also eases understanding the code.

I removed the Dial() method from the Client interface so that
TCP-based transports can manage the connections on behalf
of the application.  In particualr, the implementations of these
interfaces can transparently reconnect to the server in case the
server closed the connection on us.

Finally, I merged the getips, getnameservers, and dnssd commands
into the dnsclient tool, exposing each as a speical "meta query"
type  (see the usgae statement, and specifically th meta query
value that may be passed to the -type option.)